refactor: navigation
-Move navigation folder into components folder.
