+++ client/src/components/navigation/__tests__/NavBar.test.tsx
